[PRODUCT-260] Move connector types to separate package (#377)
* Makes sure PR number is part of cdn url

* try to get rid of the extra /merge in path

* Use a better ENV VAR to get PR id

* Change PR Message to show extra data

* temp enable pr comment on change

* Final tweaks

* Update to working azure CDN links

* Revert to normal PR comment operation

* Add connectors package
* building blocks of connectors
* publish to GH npm on main merge
* publish to npm on release
* do not use enums in type definition
 => tsc does not translate these enums to strings

* Integrates SDK latest changes

* Also build connectors when  monorepo build

* Add build commands to package.json

* Bring MediaConnector definition up to date

* Add Docs

* Fix typings not conform to studio engine
