+++ .github/workflows/pr-comment-on-open.yml
@@ -1,4 +1,4 @@
-name: Comment deploy URL on PR open
+name: PR Comment
 on:
     pull_request:
         types: [opened]
