+++ .github/workflows/publish-docs.yml
@@ -1,5 +1,5 @@
 # Simple workflow for deploying static content to GitHub Pages
-name: Deploy static content to Pages
+name: Publish Docs
 
 on:
     # Runs when released
