+++ packages/connectors/.prettierrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+    semi: true,
+    trailingComma: 'all',
+    jsxSingleQuote: false,
+    singleQuote: true,
+    printWidth: 120,
+    tabWidth: 4,
+    endOfLine: 'auto',
+};
