+++ packages/connectors/src/index.ts
@@ -1,0 +1,3 @@
+export * as Connector from './Connector.Shared';
+export * as Font from './FontConnector';
+export * as Media from './MediaConnector';
